crs/src/crs-cp-user/preprocessor/requirements.txt
GitPython
tree-sitter
tree-sitter-c
pygments
langchain
python-dotenv
litellm
openai
langchain_community==0.2.4
